+++ src/app/layout.js
@@ -1,5 +1,4 @@
 import localFont from 'next/font/local';
-import Sidebar from './components/Sidebar'; // Import the Sidebar component
 import './globals.css';
 
 const geistSans = localFont({
@@ -24,8 +23,6 @@ export default function RootLayout({ children }) {
     <html lang="en" data-theme="light">
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
         <div className="flex">
-          {/* Conditionally render Sidebar based on authentication */}
-          <Sidebar />
 
           <div className="flex-1 p-4">
             {/* Page Content */}
