+++ src/app/page.js
@@ -1,86 +1,138 @@
+// Dashboard.js
 'use client';
 
-import { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { useRouter } from 'next/navigation';
-import Link from 'next/link';
+import PieChart from './components/PieChart';
+import ProjectsTable from './components/ProjectsTable';
+import Filters from './components/Filter';
+import { filterProjects, sortProjects } from './utils/utils';
+import Sidebar from './components/Sidebar';
 
-const Home = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+const Dashboard = () => {
+  const [projects, setProjects] = useState([]);
+  const [loading, setLoading] = useState(true);
   const [error, setError] = useState('');
+  const [filters, setFilters] = useState({
+    name: '',
+    company: '',
+    status: '',
+    startDate: '',
+    endDate: '',
+  });
+  const [sortedColumn, setSortedColumn] = useState('');
+  const [sortOrder, setSortOrder] = useState('asc');
   const router = useRouter();
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
+  useEffect(() => {
+    const fetchProjects = async () => {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        router.push('/login');
+        return;
+      }
+
+      try {
+        const response = await axios.get(
+          `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/projects`,
+          {
+            headers: {
+              'Content-Type': 'application/json',
+              Accept: 'application/json',
+              Authorization: `Bearer ${token}`,
+            },
+          }
+        );
+        setProjects(response.data);
+      } catch (error) {
+        console.error('Error fetching projects:', error);
+        setProjects([]);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchProjects();
+  }, [router]);
+
+  const handleLogout = async () => {
+    const token = localStorage.getItem('token');
 
     try {
-      const response = await axios.post(
-        `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/sign_in`,
-        {
-          user: {
-            email,
-            password,
-          },
+      await axios.delete(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/sign_out`, {
+        headers: {
+          'Content-Type': 'application/json',
+          Accept: 'application/json',
+          Authorization: `Bearer ${token}`,
         },
-        {
-          headers: {
-            'Content-Type': 'application/json',
-            'Accept': 'application/json',
-          },
-        }
-      );
-
-      localStorage.setItem('token', response.data.token);
-      router.push('/dashboard');
+      });
+      localStorage.removeItem('token');
+      router.push('/login');
     } catch (err) {
       setError(err.response.data.message);
     }
   };
 
+  const handleFilterChange = (e) => {
+    setFilters({ ...filters, [e.target.name]: e.target.value });
+  };
+
+  const filteredProjects = filterProjects(projects, filters);
+  const sortedProjects = sortProjects(filteredProjects, sortedColumn, sortOrder);
+
+  const handleSort = (column) => {
+    const newSortOrder = sortedColumn === column && sortOrder === 'asc' ? 'desc' : 'asc';
+    setSortedColumn(column);
+    setSortOrder(newSortOrder);
+  };
+
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center min-h-screen">
+        <div className="text-xl text-gray-600">Loading...</div>
+      </div>
+    );
+  }
+
   return (
-    <div className="flex justify-center items-center min-h-screen bg-gray-100 p-4">
-      <div className="w-full max-w-md bg-white rounded-lg shadow-lg p-8">
-        <h1 className="text-2xl font-semibold text-center mb-6">Login</h1>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label htmlFor="email" className="block text-sm font-medium">Email</label>
-            <input
-              id="email"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-              className="mt-2 block w-full px-4 py-2 border border-gray-300 rounded-md"
-            />
-          </div>
+    <div className="flex min-h-screen">
+      <Sidebar />
 
-          <div>
-            <label htmlFor="password" className="block text-sm font-medium">Password</label>
-            <input
-              id="password"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-              className="mt-2 block w-full px-4 py-2 border border-gray-300 rounded-md"
-            />
-          </div>
+      <div className="flex-1 bg-gray-100 text-black">
+        <div className="max-w-7xl mx-auto p-4">
+          <header className="flex justify-between items-center mb-8">
+            <h1 className="text-4xl font-bold">Project Management Dashboard</h1>
+            <button
+              onClick={handleLogout}
+              className="bg-blue-500 text-white px-6 py-2 rounded-md shadow-lg hover:bg-blue-600 transition duration-300"
+            >
+              Logout
+            </button>
+          </header>
 
-          <button
-            type="submit"
-            className="w-full py-2 px-4 bg-blue-500 text-white rounded-md hover:bg-blue-600"
-          >
-            Login
-          </button>
-        </form>
-        <p className="mt-4 text-center">
-          Don't have an account? <Link href="/signup">Sign Up</Link>
-        </p>
-        {error && <p className="mt-4 text-sm text-red-500 text-center">{error}</p>}
+          {error && <div className="text-red-500">{error}</div>}
+
+          {/* Graph Section */}
+          <section className="mb-8">
+            <h2 className="text-2xl font-semibold mb-4">Project Status Overview</h2>
+            <PieChart projects={projects} />
+          </section>
+
+          {/* Filters Section */}
+          <Filters filters={filters} handleFilterChange={handleFilterChange} />
+
+          {/* Projects Table */}
+          <ProjectsTable
+            projects={sortedProjects}
+            sortedColumn={sortedColumn}
+            sortOrder={sortOrder}
+            handleSort={handleSort}
+          />
+        </div>
       </div>
     </div>
   );
 };
 
-export default Home;
+export default Dashboard;
